requirements.txt
h5py==2.9.0
numpy==1.16.1
pandas==0.24.1
matplotlib
opencv-python

scikit-learn
scipy==1.1.0

scikit-image

Pillow
sklearn